Files up until compare.rb
