+++ lib/bringit.rb
@@ -11,33 +11,33 @@ require 'active_support/core_ext/module/delegation'
 require 'rugged'
 require "charlock_holmes"
 
-# Gitlab::Git
-require_relative "gitlab_git/popen"
-require_relative 'gitlab_git/encoding_helper'
-require_relative 'gitlab_git/path_helper'
-require_relative "gitlab_git/blame"
-require_relative "gitlab_git/blob"
-require_relative "gitlab_git/commit"
-require_relative "gitlab_git/commit_stats"
-require_relative "gitlab_git/compare"
-require_relative "gitlab_git/diff"
-require_relative "gitlab_git/diff_collection"
-require_relative "gitlab_git/hook"
-require_relative "gitlab_git/index"
-require_relative "gitlab_git/rev_list"
-require_relative "gitlab_git/repository"
-require_relative "gitlab_git/tree"
-require_relative "gitlab_git/blob_snippet"
-require_relative "gitlab_git/ref"
-require_relative "gitlab_git/branch"
-require_relative "gitlab_git/tag"
-require_relative "gitlab_git/util"
-require_relative "gitlab_git/attributes"
-require_relative "gitlab_git/version_info"
-require_relative "gitlab_git/committing"
-require_relative "gitlab_git/cloning"
-require_relative "gitlab_git/pulling"
-require_relative "gitlab_git/wrapper"
+# Bringit
+require_relative "bringit/popen"
+require_relative 'bringit/encoding_helper'
+require_relative 'bringit/path_helper'
+require_relative "bringit/blame"
+require_relative "bringit/blob"
+require_relative "bringit/commit"
+require_relative "bringit/commit_stats"
+require_relative "bringit/compare"
+require_relative "bringit/diff"
+require_relative "bringit/diff_collection"
+require_relative "bringit/hook"
+require_relative "bringit/index"
+require_relative "bringit/rev_list"
+require_relative "bringit/repository"
+require_relative "bringit/tree"
+require_relative "bringit/blob_snippet"
+require_relative "bringit/ref"
+require_relative "bringit/branch"
+require_relative "bringit/tag"
+require_relative "bringit/util"
+require_relative "bringit/attributes"
+require_relative "bringit/version_info"
+require_relative "bringit/committing"
+require_relative "bringit/cloning"
+require_relative "bringit/pulling"
+require_relative "bringit/wrapper"
 
 module Gitlab
   module Git
