+++ lib/bringit/diff.rb
