+++ lib/bringit/diff_collection.rb
