+++ lib/bringit/encoding_helper.rb
