+++ lib/bringit/hook.rb
