+++ lib/bringit/path_helper.rb
