+++ lib/bringit/rev_list.rb
