+++ lib/bringit/tree.rb
