+++ lib/bringit/util.rb
