+++ lib/bringit/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Bringit
+  VERSION = '1.0.0'
+end
