+++ lib/bringit/version_info.rb
