--- lib/gitlab_git/branch.rb
@@ -1,9 +1,0 @@
-module Gitlab
-  module Git
-    class Branch < Ref
-      def self.find(repository,name)
-        repository.branches.find { |branch| branch.name == name }
-      end
-    end
-  end
-end
