--- lib/gitlab_git/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Gitlab
-  module Git
-    VERSION = '11.0.0'
-  end
-end
